requirements.txt
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
xgboost>=1.7.0
optuna>=3.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
kaggle>=1.5.16
scipy>=1.7.0
PyYAML>=6.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
pytest>=7.0.0 